backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
sendgrid==6.10.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosqlite==0.19.0